Add styles for feed, rename profileform component file
