+++ src/assets/scss/App.scss
@@ -468,4 +468,24 @@ h1, h2, h3, h4 {
 
 .add-item {
   font-size: 13px;
+  cursor: pointer;
+}
+
+.ant-list-item {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  font-size: 12px;
+}
+
+.feed-profile-header {
+  display: flex;
+  align-items: center;
+  border-bottom: 1px solid #e8e8e8;
+  padding-bottom: 10px;
+
+  span.nickname {
+    margin-left: 10px;
+    font-weight: 700;
+  }
 }
